feat: add tablet conf and refact cluster_info
