+++ src/sdk/sql_sdk_test.h
@@ -238,14 +238,14 @@ void SQLSDKTest::CreateProcedure(hybridse::sqlcase::SqlCase& sql_case,  // NOLIN
     }
     ASSERT_NO_FATAL_FAILURE(ASSERT_TRUE(0 == status.code)) << status.msg;
     auto sp_info = router->ShowProcedure(sql_case.db(), sql_case.sp_name_, &status);
-    for (int try_n = 0; try_n < 3; try_n++) {
+    for (int try_n = 0; try_n < 5; try_n++) {
         if (sp_info && status.code == 0) {
             break;
         }
         ASSERT_TRUE(router->RefreshCatalog());
         sp_info = router->ShowProcedure(sql_case.db(), sql_case.sp_name_, &status);
         LOG(WARNING) << "Procedure not found, try " << try_n << " times";
-        sleep(1);
+        sleep(try_n + 1);
     }
     ASSERT_TRUE(sp_info && status.code == 0) << status.msg;
     if (is_batch) {
